refactor common robot ai code
