+++ common/buildcraft/robotics/boards/BoardRobotFarmer.java
@@ -23,9 +23,8 @@ import buildcraft.api.robots.ResourceIdBlock;
 import buildcraft.core.lib.inventory.filters.IStackFilter;
 import buildcraft.core.lib.utils.IBlockFilter;
 import buildcraft.robotics.ai.AIRobotFetchAndEquipItemStack;
-import buildcraft.robotics.ai.AIRobotGotoBlock;
 import buildcraft.robotics.ai.AIRobotGotoSleep;
-import buildcraft.robotics.ai.AIRobotSearchBlock;
+import buildcraft.robotics.ai.AIRobotSearchAndGotoBlock;
 import buildcraft.robotics.ai.AIRobotUseToolOnBlock;
 
 public class BoardRobotFarmer extends RedstoneBoardRobot {
@@ -52,7 +51,7 @@ public class BoardRobotFarmer extends RedstoneBoardRobot {
 				}
 			}));
 		} else {
-			startDelegateAI(new AIRobotSearchBlock(robot, new IBlockFilter() {
+			startDelegateAI(new AIRobotSearchAndGotoBlock(robot, false, new IBlockFilter() {
 				@Override
 				public boolean matches(World world, int x, int y, int z) {
 					return isDirt.get(world, x, y, z)
@@ -65,21 +64,13 @@ public class BoardRobotFarmer extends RedstoneBoardRobot {
 
 	@Override
 	public void delegateAIEnded(AIRobot ai) {
-		if (ai instanceof AIRobotSearchBlock) {
-			if (!ai.success()) {
-				startDelegateAI(new AIRobotGotoSleep(robot));
+		if (ai instanceof AIRobotSearchAndGotoBlock) {
+			if (ai.success()) {
+				blockFound = ((AIRobotSearchAndGotoBlock) ai).getBlockFound();
+				startDelegateAI(new AIRobotUseToolOnBlock(robot, blockFound));
 			} else {
-				releaseBlockFound();
-				AIRobotSearchBlock searchAI = (AIRobotSearchBlock) ai;
-				if (searchAI.takeResource()) {
-					blockFound = searchAI.blockFound;
-					startDelegateAI(new AIRobotGotoBlock(robot, searchAI.path));
-				} else {
-					startDelegateAI(new AIRobotGotoSleep(robot));
-				}
+				startDelegateAI(new AIRobotGotoSleep(robot));
 			}
-		} else if (ai instanceof AIRobotGotoBlock) {
-			startDelegateAI(new AIRobotUseToolOnBlock(robot, blockFound));
 		} else if (ai instanceof AIRobotFetchAndEquipItemStack) {
 			if (robot.getHeldItem() == null) {
 				startDelegateAI(new AIRobotGotoSleep(robot));
