+++ common/buildcraft/robotics/boards/BoardRobotPump.java
@@ -30,11 +30,10 @@ import buildcraft.api.statements.IStatementParameter;
 import buildcraft.api.statements.StatementParameterItemStack;
 import buildcraft.api.statements.StatementSlot;
 import buildcraft.core.lib.utils.IBlockFilter;
-import buildcraft.robotics.ai.AIRobotGotoBlock;
 import buildcraft.robotics.ai.AIRobotGotoSleep;
 import buildcraft.robotics.ai.AIRobotGotoStationAndUnloadFluids;
 import buildcraft.robotics.ai.AIRobotPumpBlock;
-import buildcraft.robotics.ai.AIRobotSearchBlock;
+import buildcraft.robotics.ai.AIRobotSearchAndGotoBlock;
 import buildcraft.robotics.statements.ActionRobotFilter;
 import buildcraft.transport.gates.ActionIterator;
 
@@ -62,7 +61,7 @@ public class BoardRobotPump extends RedstoneBoardRobot {
 		} else {
 			updateFilter();
 
-			startDelegateAI(new AIRobotSearchBlock(robot, new IBlockFilter() {
+			startDelegateAI(new AIRobotSearchAndGotoBlock(robot, false, new IBlockFilter() {
 
 				@Override
 				public boolean matches(World world, int x, int y, int z) {
@@ -79,24 +78,12 @@ public class BoardRobotPump extends RedstoneBoardRobot {
 
 	@Override
 	public void delegateAIEnded(AIRobot ai) {
-		if (ai instanceof AIRobotSearchBlock) {
-			if (!ai.success()) {
-				startDelegateAI(new AIRobotGotoSleep(robot));
-			} else {
-				releaseBlockFound();
-				AIRobotSearchBlock searchAI = (AIRobotSearchBlock) ai;
-				if (searchAI.takeResource()) {
-					blockFound = searchAI.blockFound;
-					startDelegateAI(new AIRobotGotoBlock(robot, searchAI.path));
-				} else {
-					startDelegateAI(new AIRobotGotoSleep(robot));
-				}
-			}
-		} else if (ai instanceof AIRobotGotoBlock) {
-			if (!ai.success()) {
-				startDelegateAI(new AIRobotGotoSleep(robot));
-			} else {
+		if (ai instanceof AIRobotSearchAndGotoBlock) {
+			if (ai.success()) {
+				blockFound = ((AIRobotSearchAndGotoBlock) ai).getBlockFound();
 				startDelegateAI(new AIRobotPumpBlock(robot, blockFound));
+			} else {
+				startDelegateAI(new AIRobotGotoSleep(robot));
 			}
 		} else if (ai instanceof AIRobotGotoStationAndUnloadFluids) {
 			releaseBlockFound();
